CMake configuration update + code style improvements
* Removed custom targets from common.cmake
* Set minimal required CMake version to 3.0
* Enabled almost all gcc warnings and fixed them in the code.
* Now project is compatible with the new QtCreator 4xx code model builder.

+++ Print/include/RootPrintSource.h
@@ -147,7 +147,7 @@ public:
     using PlotOptions = typename root_ext::HistogramPlotter<Histogram, ValueType>::Options;
     using Entry = typename root_ext::HistogramPlotter<Histogram, ValueType>::Entry;
 
-    static PlotOptions& GetDefaultPlotOptions(unsigned n)
+    static PlotOptions& GetDefaultPlotOptions(size_t n)
     {
         static std::vector<PlotOptions> options;
         if(!options.size())
