+++ Print/source/Print_TreeBranch.cxx
@@ -17,7 +17,7 @@ protected:
         const std::string name = branchName + "_hist";
         const std::string command = branchName + ">>+" + name;
 
-        TH1D* histogram = new TH1D(name.c_str(), name.c_str(), nBins, xRange.min, xRange.max);
+        TH1D* histogram = new TH1D(name.c_str(), name.c_str(), static_cast<int>(nBins), xRange.min, xRange.max);
         tree->Draw(command.c_str(), "");
         return histogram;
     }
